requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.19.0
anyio==4.3.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.3.3
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
jiter==0.7.0
multidict==6.0.4
openai==1.55.0
packaging==23.2
pydantic==2.6.3
pydantic_core==2.16.3
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.28.2
sniffio==1.3.1
telebot==0.0.5
tokenizers==0.15.2
tqdm==4.65.0
typing_extensions==4.12.2
urllib3==1.26.15
yarl==1.8.2
